Add unit tests for python_archive and support
Verify that extraction to a specified directory works as expected, as well as
some error handling (e.g. destination not writable). Also test that different
inputs result in the output dir being overwritten and the par's output changes
accordingly. Also verify the payload main is not run when the PAR_EXTRACT_ONLY
environment variable is set.